deployment/yolov5-fast-api/requirements.txt
anyio==3.6.2
click==8.1.3
fastapi==0.95.0
h11==0.14.0
httptools==0.5.0
idna==3.4
Jinja2==3.1.2
MarkupSafe==2.1.2
numpy==1.24.2
opencv-python-headless==4.7.0.72
Pillow==9.4.0
pydantic==1.10.6
python-dotenv==1.0.0
python-multipart==0.0.6
PyYAML==6.0
sniffio==1.3.0
starlette==0.26.1
typing_extensions==4.5.0
uvicorn==0.21.1
uvloop==0.17.0
watchfiles==0.18.1
websockets==10.4
